pyproject.toml
[project]
name = "rag-blueprint"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-mock>=3.14.0",
    "python-on-whales>=0.76.1",
]

[project.optional-dependencies]
utils = [
    "pre-commit>=4.1.0",
    "pydantic>=2.9.2",
    "pydantic-core>=2.23.4",
    "pydantic-settings>=2.2.1",
]
extraction = [
    "llama-index-core>=0.12.24.post1",
    "tqdm>=4.67.1",
    "markitdown[pdf]>=0.1.1",
    # Confluence
    "atlassian-python-api>=3.41.19",
    # Notion
    "more-itertools>=10.6.0",
    "notion-exporter>=0.1.1",
    "notion-client>=2.3.0",
    "notion>=0.0.28",
]
embedding = [
    "chromadb>=0.6.3",
    "llama-index-embeddings-huggingface>=0.5.2",
    "llama-index-embeddings-openai>=0.3.1",
    "llama-index-embeddings-voyageai>=0.3.5",
    "llama-index-vector-stores-chroma>=0.4.1",
    "llama-index-vector-stores-postgres>=0.4.2",
    "llama-index-vector-stores-qdrant>=0.4.3",
    "psycopg2-binary>=2.9.10",
    "transformers>=4.49.0",
]
augmentation = [
    "chainlit>=2.3.0",
    "langfuse>=2.60.2",
    "llama-index-callbacks-langfuse>=0.3.0",
    "llama-index-llms-openai>=0.3.25",
    "llama-index-llms-openai-like>=0.3.4",
    "llama-index-postprocessor-colbert-rerank>=0.3.0",
]
evaluation = ["ragas==0.1.14"]

all = [
    "rag-blueprint[utils]",
    "rag-blueprint[extraction]",
    "rag-blueprint[embedding]",
    "rag-blueprint[augmentation]",
    "rag-blueprint[evaluation]",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/"]
